requirements.txt
fastapi==0.67.0
uvicorn==0.14.0
pytest==6.2.4
pyjwt==2.1.0
pyjwt[crypto]==2.1.0
python-dotenv==0.19.0
pytest-cov
requests
git+https://github.com/Maua-Dev/models-devmaua